+++ src/MailgunAdapter.php
@@ -46,6 +46,11 @@ class MailgunAdapter extends BaseTransportAdapter
      */
     public $apiKey;
 
+    /**
+     * @var string The API endpoint that should be used
+     */
+    public $endpoint;
+
     // Public Methods
     // =========================================================================
 
@@ -57,6 +62,7 @@ class MailgunAdapter extends BaseTransportAdapter
         return [
             'apiKey' => Craft::t('mailgun', 'API Key'),
             'domain' => Craft::t('mailgun', 'Domain'),
+            'endpoint' => Craft::t('mailgun', 'Endpoint'),
         ];
     }
 
@@ -67,6 +73,7 @@ class MailgunAdapter extends BaseTransportAdapter
     {
         return [
             [['apiKey', 'domain'], 'required'],
+            [['endpoint'], 'url', 'defaultScheme' => 'https'],
         ];
     }
 
@@ -90,6 +97,10 @@ class MailgunAdapter extends BaseTransportAdapter
         $httpClientConfigurator = (new HttpClientConfigurator())
             ->setHttpClient($client)
             ->setApiKey($this->apiKey);
+
+        if ($this->endpoint) {
+            $httpClientConfigurator->setEndpoint($this->endpoint);
+        }
 
         return [
             'class' => MailgunTransport::class,